Move routes.ts out of utils directory.
